Create testing methods.

+++ DrawGame/Model/Pixel.swift
@@ -16,7 +16,10 @@ extension Pixel {
 
 extension Pixel {
     var pixelData: PixelData {
-        return PixelData(a: 255, r: r, g: g, b: b)
+        return PixelData(a: 255,
+                         r: UInt8(r),
+                         g: UInt8(g),
+                         b: UInt8(b))
     }
 }
